alpha=0.75 and robust points
After some reading, I decided to use 75% of the data instead of the
default 50% in the algorithm.  In addition, I realized that the program
outputs 2 sets of points. The “best” set is the points used in the last
step of the MCD; while the points with raw.wt==1 are the points used in
the actual calculation of the MCD covariance.  For details see Rousseau
1999